README now reflects current status of the project
Fixed 'unused parameter' warnings in 'table' module.
Fixed 'narrowing warnings' in 'table' module.
Fixed narrowing in format string in 'compiler' module.

+++ src/lib/compiler.c
@@ -35,7 +35,7 @@ static void errorAt(Parser *parser, Token *token, const char *message) {
     } else if (token->type == TOKEN_ERROR) {
         // Nothing
     } else {
-        fprintf(stderr, " at '%.*s'", token->length, token->start);
+        fprintf(stderr, " at '%.*s'", (int)token->length, token->start);
     }
 
     fprintf(stderr, ": %s\n", message);
